+++ pyfragment/methods/__init__.py
@@ -1,5 +1,5 @@
-from pyfragment.methods.premium import gift_premium
-from pyfragment.methods.stars import gift_stars
+from pyfragment.methods.premium import purchase_premium
+from pyfragment.methods.stars import purchase_stars
 from pyfragment.methods.ton import topup_ton
 
-__all__ = ["gift_premium", "gift_stars", "topup_ton"]
+__all__ = ["purchase_premium", "purchase_stars", "topup_ton"]